nit: move History to separate file
