+++ packages/utils/src/useMethods.ts
@@ -1,7 +1,8 @@
 // https://github.com/pelotom/use-methods
-import produce, { applyPatches, Patch } from "immer";
+import produce, { applyPatches } from "immer";
 import { useMemo, useEffect, useRef, useReducer, useCallback } from "react";
 import isEqualWith from "lodash.isequalwith";
+import { History } from "./History";
 
 export type SubscriberAndCallbacksFor<
   M extends MethodsOrOptions,
@@ -114,10 +115,7 @@ export function useMethods<
   initialState: any,
   queryMethods?: Q
 ): SubscriberAndCallbacksFor<MethodsOrOptions<S, R>, Q> {
-  const history = useMemo(() => {
-    console.log("Creating history");
-    return new History();
-  }, []);
+  const history = useMemo(() => new History(), []);
 
   const [reducer, methodsFactory] = useMemo(() => {
     let methods: Methods<S, R>;
@@ -246,75 +244,6 @@ export function createQuery<Q extends QueryMethods>(
     canUndo: () => history.canUndo(),
     canRedo: () => history.canRedo(),
   };
-}
-
-type Timeline = Array<{
-  patches: Patch[];
-  inversePatches: Patch[];
-}>;
-
-class History {
-  timeline: Timeline = [];
-  pointer = -1;
-  lastChange;
-
-  add(patches, inversePatches, action) {
-    if (patches.length == 0 && inversePatches.length == 0) return;
-
-    if (this.canUndo()) {
-      const { patches: currPatches } = this.timeline[this.pointer];
-
-      const now = new Date();
-      const diff = (now.getTime() - this.lastChange.getTime()) / 1000;
-
-      // Ignore similar changes that occurs within 2 seconds
-      if (diff < 2 && currPatches.length == patches.length) {
-        const isSimilar = currPatches.every((currPatch, i) => {
-          const { op: currOp, path: currPath } = currPatch;
-          const { op, path } = patches[i];
-
-          if (op == currOp && isEqualWith(path, currPath)) return true;
-          return false;
-        });
-
-        if (isSimilar) {
-          return;
-        }
-      }
-    }
-
-    this.pointer = this.pointer + 1;
-    this.timeline.length = this.pointer;
-    this.timeline[this.pointer] = { patches, inversePatches };
-
-    this.lastChange = new Date();
-  }
-
-  canUndo() {
-    return this.pointer >= 0;
-  }
-
-  canRedo() {
-    return this.pointer != this.timeline.length - 1;
-  }
-
-  undo(state) {
-    if (this.canUndo()) {
-      const { inversePatches } = this.timeline[this.pointer];
-      this.pointer = this.pointer - 1;
-      const applied = applyPatches(state, inversePatches);
-      return applied;
-    }
-  }
-
-  redo(state) {
-    if (this.canRedo()) {
-      this.pointer = this.pointer + 1;
-      const { patches } = this.timeline[this.pointer];
-      const applied = applyPatches(state, patches);
-      return applied;
-    }
-  }
 }
 
 class Watcher<S> {
